adding router navigation and rest of the pages also components header and foter
