+++ danielmoralesbweb/src/pages/Home.js
@@ -1,0 +1,15 @@
+import Header from "../components/Header";
+import Footer from "../components/Footer";
+
+function Home() {
+    return(
+        <div>
+            <div className="container">
+                <Header />
+            </div>
+            <Footer />
+        </div>
+    )
+}
+
+export default Home;
