+++ danielmoralesbweb/src/pages/Websites.js
@@ -1,0 +1,16 @@
+import Header from "../components/Header";
+import Footer from "../components/Footer";
+
+function NoPage() {
+    return(
+        <div>
+            <div className="container">
+                <Header />
+                <p>Websites</p>
+            </div>
+            <Footer />
+        </div>
+    )
+}
+
+export default NoPage;
